Safari-611.1.10/LayoutTests/media/fullscreen-api-enabled-media-with-presentation-mode-expected.txt

Tests that the video's presentation mode is updated accordingly after entering and exiting Full Screen

EXPECTED (video.webkitPresentationMode == 'inline') OK
Going into Full Screen
EXPECTED (video.webkitPresentationMode == 'fullscreen') OK
Exiting Full Screen
EXPECTED (video.webkitPresentationMode == 'inline') OK
END OF TEST
